requirements.txt
Flask==2.3.5
openai==1.104.2
gspread==5.8.2
oauth2client==4.1.3
python-dotenv==1.0.1
